hardware/nubia: fingerprint: protect session callback using mutex
Message handling involves calling the session callback. If setting
session callback is unprotected, it's possible for session callback
to be set to nullptr while a message is being handled.

This happens in real life when authenticate future is canceled and then
the session is closed, setting callback to nullptr, after which the
cancel error message is injected to stop the current authenticate
operation, and based on documentation of ISession, the cancel error
should also be sent to the framework, but since session callback has
been set to null, callback is no longer available to do that.

Protect all usages of session callback using the same mutex.

Change-Id: Ib1e9cc6bc5785a55b349d511ef04da534780bea5